requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.10
importlib_metadata==8.5.0
packaging==24.1
PyJWT==2.9.0
python-arango==8.1.2
pytz==2024.2
requests==2.32.3
requests-toolbelt==1.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
